% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoefMod.R
\name{CoefMod}
\alias{CoefMod}
\title{Coefficient Modulo}
\usage{
CoefMod(x, k)
}
\arguments{
\item{x}{polynomial from the polynom package}

\item{k}{the modulo}
}
\value{
polynomial of the polynom class
}
\description{
Coefficient Modulo
}
\examples{
polynomial = polynomial(c(5, 3, 6))
print(polynomial)

CoefMod(polynomial, 5)
}
