% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenDiscrGauss.R
\name{GenDiscrGauss}
\alias{GenDiscrGauss}
\title{Generate Polynomial with Discrete Gaussian Coefficients}
\usage{
GenDiscrGauss(n, s = 3)
}
\arguments{
\item{n}{the order}

\item{s}{scale the sigma (down)}
}
\value{
polynomial of the form x^^n + 1
}
\description{
Generate Polynomial with Discrete Gaussian Coefficients
}
\examples{
n = 5
GenDiscrGauss(n)

GenDiscrGauss(n=5, s=2)
}
