% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenPolyMod.R
\name{GenPolyMod}
\alias{GenPolyMod}
\title{Generate Polynomial Modulo}
\usage{
GenPolyMod(n)
}
\arguments{
\item{n}{the order}
}
\value{
polynomial of the form x^^n + 1
}
\description{
Generate Polynomial Modulo
}
\examples{
n = 5
GenPolyMod(5)
}
