% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenTernary.R
\name{GenTernary}
\alias{GenTernary}
\title{Generate Polynomial with Ternary}
\usage{
GenTernary(n)
}
\arguments{
\item{n}{the order}
}
\value{
ternary polynomial of order x^^n with coefficients (-1,0,1)
}
\description{
Generate Polynomial with Ternary
}
\examples{
n = 5
GenTernary(n)
}
