% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenUnif.R
\name{GenUnif}
\alias{GenUnif}
\title{Generate Polynomial with Uniform Distribution Coefficients}
\usage{
GenUnif(n, q)
}
\arguments{
\item{n}{the order}

\item{q}{the ciphermod of coefficients}
}
\value{
polynomial of order x^^n with coefficients 0,..,q
}
\description{
Generate Polynomial with Uniform Distribution Coefficients
}
\examples{
n = 5
q = 7
GenUnif(n, q)
}
